8252324: Signal related code should be shared among POSIX platforms
