+++ src/hotspot/os/aix/os_aix.hpp
@@ -34,12 +34,6 @@ static bool zero_page_read_protected() { return false; }
 class Aix {
   friend class os;
 
-  static bool libjsig_is_loaded;        // libjsig that interposes sigaction(),
-                                        // __sigaction(), signal() is loaded
-  static struct sigaction *(*get_signal_action)(int);
-
-  static void check_signal_handler(int sig);
-
  private:
 
   static julong _physical_memory;
@@ -89,7 +83,6 @@ class Aix {
  public:
   static void init_thread_fpu_state();
   static pthread_t main_thread(void)                                { return _main_thread; }
-  static void hotspot_sigmask(Thread* thread);
 
   // Given an address, returns the size of the page backing that address
   static size_t query_pagesize(void* p);
@@ -106,23 +99,6 @@ class Aix {
   static void ucontext_set_pc(ucontext_t* uc, address pc);
 
   static bool get_frame_at_stack_banging_point(JavaThread* thread, ucontext_t* uc, frame* fr);
-
-  // This boolean allows users to forward their own non-matching signals
-  // to JVM_handle_aix_signal, harmlessly.
-  static bool signal_handlers_are_installed;
-
-  static int get_our_sigflags(int);
-  static void set_our_sigflags(int, int);
-  static void signal_sets_init();
-  static void install_signal_handlers();
-  static void set_signal_handler(int, bool);
-
-  static sigset_t* unblocked_signals();
-  static sigset_t* vm_signals();
-
-  // For signal-chaining
-  static struct sigaction *get_chained_signal_action(int sig);
-  static bool chained_handler(int sig, siginfo_t* siginfo, void* context);
 
   // libpthread version string
   static void libpthread_init();
@@ -204,7 +180,6 @@ class Aix {
   // (on AIX, using libperfstat, on PASE with libo4.so).
   // Returns true if ok, false if error.
   static bool get_meminfo(meminfo_t* pmi);
-
 };
 
 #endif // OS_AIX_OS_AIX_HPP
