+++ src/hotspot/os/posix/jvm_posix.cpp
@@ -26,6 +26,7 @@
 #include "jvm.h"
 #include "runtime/interfaceSupport.inline.hpp"
 #include "runtime/osThread.hpp"
+#include "signals_posix.hpp"
 
 #include <signal.h>
 
@@ -78,7 +79,7 @@ JVM_ENTRY_NO_ENV(void*, JVM_RegisterSignal(jint sig, void* handler))
     case SHUTDOWN2_SIGNAL:
     case SHUTDOWN3_SIGNAL:
       if (ReduceSignalUsage) return (void*)-1;
-      if (os::Posix::is_sig_ignored(sig)) return (void*)1;
+      if (PosixSignals::is_sig_ignored(sig)) return (void*)1;
   }
 
   void* oldHandler = os::signal(sig, newHandler);
@@ -102,7 +103,7 @@ JVM_ENTRY_NO_ENV(jboolean, JVM_RaiseSignal(jint sig))
     }
   }
   else if ((sig == SHUTDOWN1_SIGNAL || sig == SHUTDOWN2_SIGNAL ||
-            sig == SHUTDOWN3_SIGNAL) && os::Posix::is_sig_ignored(sig)) {
+            sig == SHUTDOWN3_SIGNAL) && PosixSignals::is_sig_ignored(sig)) {
     // do not allow SHUTDOWN1_SIGNAL to be raised when SHUTDOWN1_SIGNAL
     // is ignored, since no handler for them is actually registered in JVM
     // or via JVM_RegisterSignal.
