+++ src/hotspot/os/posix/os_posix.cpp
@@ -53,13 +53,6 @@
 #include <unistd.h>
 #include <utmpx.h>
 
-// Todo: provide a os::get_max_process_id() or similar. Number of processes
-// may have been configured, can be read more accurately from proc fs etc.
-#ifndef MAX_PID
-#define MAX_PID INT_MAX
-#endif
-#define IS_VALID_PID(p) (p > 0 && p < MAX_PID)
-
 #define ROOT_UID 0
 
 #ifndef MAP_ANONYMOUS
@@ -705,603 +698,6 @@ void os::naked_short_sleep(jlong ms) {
   return;
 }
 
-static const struct {
-  int sig; const char* name;
-}
- g_signal_info[] =
-  {
-  {  SIGABRT,     "SIGABRT" },
-#ifdef SIGAIO
-  {  SIGAIO,      "SIGAIO" },
-#endif
-  {  SIGALRM,     "SIGALRM" },
-#ifdef SIGALRM1
-  {  SIGALRM1,    "SIGALRM1" },
-#endif
-  {  SIGBUS,      "SIGBUS" },
-#ifdef SIGCANCEL
-  {  SIGCANCEL,   "SIGCANCEL" },
-#endif
-  {  SIGCHLD,     "SIGCHLD" },
-#ifdef SIGCLD
-  {  SIGCLD,      "SIGCLD" },
-#endif
-  {  SIGCONT,     "SIGCONT" },
-#ifdef SIGCPUFAIL
-  {  SIGCPUFAIL,  "SIGCPUFAIL" },
-#endif
-#ifdef SIGDANGER
-  {  SIGDANGER,   "SIGDANGER" },
-#endif
-#ifdef SIGDIL
-  {  SIGDIL,      "SIGDIL" },
-#endif
-#ifdef SIGEMT
-  {  SIGEMT,      "SIGEMT" },
-#endif
-  {  SIGFPE,      "SIGFPE" },
-#ifdef SIGFREEZE
-  {  SIGFREEZE,   "SIGFREEZE" },
-#endif
-#ifdef SIGGFAULT
-  {  SIGGFAULT,   "SIGGFAULT" },
-#endif
-#ifdef SIGGRANT
-  {  SIGGRANT,    "SIGGRANT" },
-#endif
-  {  SIGHUP,      "SIGHUP" },
-  {  SIGILL,      "SIGILL" },
-#ifdef SIGINFO
-  {  SIGINFO,     "SIGINFO" },
-#endif
-  {  SIGINT,      "SIGINT" },
-#ifdef SIGIO
-  {  SIGIO,       "SIGIO" },
-#endif
-#ifdef SIGIOINT
-  {  SIGIOINT,    "SIGIOINT" },
-#endif
-#ifdef SIGIOT
-// SIGIOT is there for BSD compatibility, but on most Unices just a
-// synonym for SIGABRT. The result should be "SIGABRT", not
-// "SIGIOT".
-#if (SIGIOT != SIGABRT )
-  {  SIGIOT,      "SIGIOT" },
-#endif
-#endif
-#ifdef SIGKAP
-  {  SIGKAP,      "SIGKAP" },
-#endif
-  {  SIGKILL,     "SIGKILL" },
-#ifdef SIGLOST
-  {  SIGLOST,     "SIGLOST" },
-#endif
-#ifdef SIGLWP
-  {  SIGLWP,      "SIGLWP" },
-#endif
-#ifdef SIGLWPTIMER
-  {  SIGLWPTIMER, "SIGLWPTIMER" },
-#endif
-#ifdef SIGMIGRATE
-  {  SIGMIGRATE,  "SIGMIGRATE" },
-#endif
-#ifdef SIGMSG
-  {  SIGMSG,      "SIGMSG" },
-#endif
-  {  SIGPIPE,     "SIGPIPE" },
-#ifdef SIGPOLL
-  {  SIGPOLL,     "SIGPOLL" },
-#endif
-#ifdef SIGPRE
-  {  SIGPRE,      "SIGPRE" },
-#endif
-  {  SIGPROF,     "SIGPROF" },
-#ifdef SIGPTY
-  {  SIGPTY,      "SIGPTY" },
-#endif
-#ifdef SIGPWR
-  {  SIGPWR,      "SIGPWR" },
-#endif
-  {  SIGQUIT,     "SIGQUIT" },
-#ifdef SIGRECONFIG
-  {  SIGRECONFIG, "SIGRECONFIG" },
-#endif
-#ifdef SIGRECOVERY
-  {  SIGRECOVERY, "SIGRECOVERY" },
-#endif
-#ifdef SIGRESERVE
-  {  SIGRESERVE,  "SIGRESERVE" },
-#endif
-#ifdef SIGRETRACT
-  {  SIGRETRACT,  "SIGRETRACT" },
-#endif
-#ifdef SIGSAK
-  {  SIGSAK,      "SIGSAK" },
-#endif
-  {  SIGSEGV,     "SIGSEGV" },
-#ifdef SIGSOUND
-  {  SIGSOUND,    "SIGSOUND" },
-#endif
-#ifdef SIGSTKFLT
-  {  SIGSTKFLT,    "SIGSTKFLT" },
-#endif
-  {  SIGSTOP,     "SIGSTOP" },
-  {  SIGSYS,      "SIGSYS" },
-#ifdef SIGSYSERROR
-  {  SIGSYSERROR, "SIGSYSERROR" },
-#endif
-#ifdef SIGTALRM
-  {  SIGTALRM,    "SIGTALRM" },
-#endif
-  {  SIGTERM,     "SIGTERM" },
-#ifdef SIGTHAW
-  {  SIGTHAW,     "SIGTHAW" },
-#endif
-  {  SIGTRAP,     "SIGTRAP" },
-#ifdef SIGTSTP
-  {  SIGTSTP,     "SIGTSTP" },
-#endif
-  {  SIGTTIN,     "SIGTTIN" },
-  {  SIGTTOU,     "SIGTTOU" },
-#ifdef SIGURG
-  {  SIGURG,      "SIGURG" },
-#endif
-  {  SIGUSR1,     "SIGUSR1" },
-  {  SIGUSR2,     "SIGUSR2" },
-#ifdef SIGVIRT
-  {  SIGVIRT,     "SIGVIRT" },
-#endif
-  {  SIGVTALRM,   "SIGVTALRM" },
-#ifdef SIGWAITING
-  {  SIGWAITING,  "SIGWAITING" },
-#endif
-#ifdef SIGWINCH
-  {  SIGWINCH,    "SIGWINCH" },
-#endif
-#ifdef SIGWINDOW
-  {  SIGWINDOW,   "SIGWINDOW" },
-#endif
-  {  SIGXCPU,     "SIGXCPU" },
-  {  SIGXFSZ,     "SIGXFSZ" },
-#ifdef SIGXRES
-  {  SIGXRES,     "SIGXRES" },
-#endif
-  { -1, NULL }
-};
-
-// Returned string is a constant. For unknown signals "UNKNOWN" is returned.
-const char* os::Posix::get_signal_name(int sig, char* out, size_t outlen) {
-
-  const char* ret = NULL;
-
-#ifdef SIGRTMIN
-  if (sig >= SIGRTMIN && sig <= SIGRTMAX) {
-    if (sig == SIGRTMIN) {
-      ret = "SIGRTMIN";
-    } else if (sig == SIGRTMAX) {
-      ret = "SIGRTMAX";
-    } else {
-      jio_snprintf(out, outlen, "SIGRTMIN+%d", sig - SIGRTMIN);
-      return out;
-    }
-  }
-#endif
-
-  if (sig > 0) {
-    for (int idx = 0; g_signal_info[idx].sig != -1; idx ++) {
-      if (g_signal_info[idx].sig == sig) {
-        ret = g_signal_info[idx].name;
-        break;
-      }
-    }
-  }
-
-  if (!ret) {
-    if (!is_valid_signal(sig)) {
-      ret = "INVALID";
-    } else {
-      ret = "UNKNOWN";
-    }
-  }
-
-  if (out && outlen > 0) {
-    strncpy(out, ret, outlen);
-    out[outlen - 1] = '\0';
-  }
-  return out;
-}
-
-int os::Posix::get_signal_number(const char* signal_name) {
-  char tmp[30];
-  const char* s = signal_name;
-  if (s[0] != 'S' || s[1] != 'I' || s[2] != 'G') {
-    jio_snprintf(tmp, sizeof(tmp), "SIG%s", signal_name);
-    s = tmp;
-  }
-  for (int idx = 0; g_signal_info[idx].sig != -1; idx ++) {
-    if (strcmp(g_signal_info[idx].name, s) == 0) {
-      return g_signal_info[idx].sig;
-    }
-  }
-  return -1;
-}
-
-int os::get_signal_number(const char* signal_name) {
-  return os::Posix::get_signal_number(signal_name);
-}
-
-// Returns true if signal number is valid.
-bool os::Posix::is_valid_signal(int sig) {
-  // MacOS not really POSIX compliant: sigaddset does not return
-  // an error for invalid signal numbers. However, MacOS does not
-  // support real time signals and simply seems to have just 33
-  // signals with no holes in the signal range.
-#ifdef __APPLE__
-  return sig >= 1 && sig < NSIG;
-#else
-  // Use sigaddset to check for signal validity.
-  sigset_t set;
-  sigemptyset(&set);
-  if (sigaddset(&set, sig) == -1 && errno == EINVAL) {
-    return false;
-  }
-  return true;
-#endif
-}
-
-bool os::Posix::is_sig_ignored(int sig) {
-  struct sigaction oact;
-  sigaction(sig, (struct sigaction*)NULL, &oact);
-  void* ohlr = oact.sa_sigaction ? CAST_FROM_FN_PTR(void*,  oact.sa_sigaction)
-                                 : CAST_FROM_FN_PTR(void*,  oact.sa_handler);
-  if (ohlr == CAST_FROM_FN_PTR(void*, SIG_IGN)) {
-    return true;
-  } else {
-    return false;
-  }
-}
-
-// Returns:
-// NULL for an invalid signal number
-// "SIG<num>" for a valid but unknown signal number
-// signal name otherwise.
-const char* os::exception_name(int sig, char* buf, size_t size) {
-  if (!os::Posix::is_valid_signal(sig)) {
-    return NULL;
-  }
-  const char* const name = os::Posix::get_signal_name(sig, buf, size);
-  if (strcmp(name, "UNKNOWN") == 0) {
-    jio_snprintf(buf, size, "SIG%d", sig);
-  }
-  return buf;
-}
-
-#define NUM_IMPORTANT_SIGS 32
-// Returns one-line short description of a signal set in a user provided buffer.
-const char* os::Posix::describe_signal_set_short(const sigset_t* set, char* buffer, size_t buf_size) {
-  assert(buf_size == (NUM_IMPORTANT_SIGS + 1), "wrong buffer size");
-  // Note: for shortness, just print out the first 32. That should
-  // cover most of the useful ones, apart from realtime signals.
-  for (int sig = 1; sig <= NUM_IMPORTANT_SIGS; sig++) {
-    const int rc = sigismember(set, sig);
-    if (rc == -1 && errno == EINVAL) {
-      buffer[sig-1] = '?';
-    } else {
-      buffer[sig-1] = rc == 0 ? '0' : '1';
-    }
-  }
-  buffer[NUM_IMPORTANT_SIGS] = 0;
-  return buffer;
-}
-
-// Prints one-line description of a signal set.
-void os::Posix::print_signal_set_short(outputStream* st, const sigset_t* set) {
-  char buf[NUM_IMPORTANT_SIGS + 1];
-  os::Posix::describe_signal_set_short(set, buf, sizeof(buf));
-  st->print("%s", buf);
-}
-
-// Writes one-line description of a combination of sigaction.sa_flags into a user
-// provided buffer. Returns that buffer.
-const char* os::Posix::describe_sa_flags(int flags, char* buffer, size_t size) {
-  char* p = buffer;
-  size_t remaining = size;
-  bool first = true;
-  int idx = 0;
-
-  assert(buffer, "invalid argument");
-
-  if (size == 0) {
-    return buffer;
-  }
-
-  strncpy(buffer, "none", size);
-
-  const struct {
-    // NB: i is an unsigned int here because SA_RESETHAND is on some
-    // systems 0x80000000, which is implicitly unsigned.  Assignining
-    // it to an int field would be an overflow in unsigned-to-signed
-    // conversion.
-    unsigned int i;
-    const char* s;
-  } flaginfo [] = {
-    { SA_NOCLDSTOP, "SA_NOCLDSTOP" },
-    { SA_ONSTACK,   "SA_ONSTACK"   },
-    { SA_RESETHAND, "SA_RESETHAND" },
-    { SA_RESTART,   "SA_RESTART"   },
-    { SA_SIGINFO,   "SA_SIGINFO"   },
-    { SA_NOCLDWAIT, "SA_NOCLDWAIT" },
-    { SA_NODEFER,   "SA_NODEFER"   },
-#ifdef AIX
-    { SA_ONSTACK,   "SA_ONSTACK"   },
-    { SA_OLDSTYLE,  "SA_OLDSTYLE"  },
-#endif
-    { 0, NULL }
-  };
-
-  for (idx = 0; flaginfo[idx].s && remaining > 1; idx++) {
-    if (flags & flaginfo[idx].i) {
-      if (first) {
-        jio_snprintf(p, remaining, "%s", flaginfo[idx].s);
-        first = false;
-      } else {
-        jio_snprintf(p, remaining, "|%s", flaginfo[idx].s);
-      }
-      const size_t len = strlen(p);
-      p += len;
-      remaining -= len;
-    }
-  }
-
-  buffer[size - 1] = '\0';
-
-  return buffer;
-}
-
-// Prints one-line description of a combination of sigaction.sa_flags.
-void os::Posix::print_sa_flags(outputStream* st, int flags) {
-  char buffer[0x100];
-  os::Posix::describe_sa_flags(flags, buffer, sizeof(buffer));
-  st->print("%s", buffer);
-}
-
-// Helper function for os::Posix::print_siginfo_...():
-// return a textual description for signal code.
-struct enum_sigcode_desc_t {
-  const char* s_name;
-  const char* s_desc;
-};
-
-static bool get_signal_code_description(const siginfo_t* si, enum_sigcode_desc_t* out) {
-
-  const struct {
-    int sig; int code; const char* s_code; const char* s_desc;
-  } t1 [] = {
-    { SIGILL,  ILL_ILLOPC,   "ILL_ILLOPC",   "Illegal opcode." },
-    { SIGILL,  ILL_ILLOPN,   "ILL_ILLOPN",   "Illegal operand." },
-    { SIGILL,  ILL_ILLADR,   "ILL_ILLADR",   "Illegal addressing mode." },
-    { SIGILL,  ILL_ILLTRP,   "ILL_ILLTRP",   "Illegal trap." },
-    { SIGILL,  ILL_PRVOPC,   "ILL_PRVOPC",   "Privileged opcode." },
-    { SIGILL,  ILL_PRVREG,   "ILL_PRVREG",   "Privileged register." },
-    { SIGILL,  ILL_COPROC,   "ILL_COPROC",   "Coprocessor error." },
-    { SIGILL,  ILL_BADSTK,   "ILL_BADSTK",   "Internal stack error." },
-#if defined(IA64) && defined(LINUX)
-    { SIGILL,  ILL_BADIADDR, "ILL_BADIADDR", "Unimplemented instruction address" },
-    { SIGILL,  ILL_BREAK,    "ILL_BREAK",    "Application Break instruction" },
-#endif
-    { SIGFPE,  FPE_INTDIV,   "FPE_INTDIV",   "Integer divide by zero." },
-    { SIGFPE,  FPE_INTOVF,   "FPE_INTOVF",   "Integer overflow." },
-    { SIGFPE,  FPE_FLTDIV,   "FPE_FLTDIV",   "Floating-point divide by zero." },
-    { SIGFPE,  FPE_FLTOVF,   "FPE_FLTOVF",   "Floating-point overflow." },
-    { SIGFPE,  FPE_FLTUND,   "FPE_FLTUND",   "Floating-point underflow." },
-    { SIGFPE,  FPE_FLTRES,   "FPE_FLTRES",   "Floating-point inexact result." },
-    { SIGFPE,  FPE_FLTINV,   "FPE_FLTINV",   "Invalid floating-point operation." },
-    { SIGFPE,  FPE_FLTSUB,   "FPE_FLTSUB",   "Subscript out of range." },
-    { SIGSEGV, SEGV_MAPERR,  "SEGV_MAPERR",  "Address not mapped to object." },
-    { SIGSEGV, SEGV_ACCERR,  "SEGV_ACCERR",  "Invalid permissions for mapped object." },
-#ifdef AIX
-    // no explanation found what keyerr would be
-    { SIGSEGV, SEGV_KEYERR,  "SEGV_KEYERR",  "key error" },
-#endif
-#if defined(IA64) && !defined(AIX)
-    { SIGSEGV, SEGV_PSTKOVF, "SEGV_PSTKOVF", "Paragraph stack overflow" },
-#endif
-    { SIGBUS,  BUS_ADRALN,   "BUS_ADRALN",   "Invalid address alignment." },
-    { SIGBUS,  BUS_ADRERR,   "BUS_ADRERR",   "Nonexistent physical address." },
-    { SIGBUS,  BUS_OBJERR,   "BUS_OBJERR",   "Object-specific hardware error." },
-    { SIGTRAP, TRAP_BRKPT,   "TRAP_BRKPT",   "Process breakpoint." },
-    { SIGTRAP, TRAP_TRACE,   "TRAP_TRACE",   "Process trace trap." },
-    { SIGCHLD, CLD_EXITED,   "CLD_EXITED",   "Child has exited." },
-    { SIGCHLD, CLD_KILLED,   "CLD_KILLED",   "Child has terminated abnormally and did not create a core file." },
-    { SIGCHLD, CLD_DUMPED,   "CLD_DUMPED",   "Child has terminated abnormally and created a core file." },
-    { SIGCHLD, CLD_TRAPPED,  "CLD_TRAPPED",  "Traced child has trapped." },
-    { SIGCHLD, CLD_STOPPED,  "CLD_STOPPED",  "Child has stopped." },
-    { SIGCHLD, CLD_CONTINUED,"CLD_CONTINUED","Stopped child has continued." },
-#ifdef SIGPOLL
-    { SIGPOLL, POLL_OUT,     "POLL_OUT",     "Output buffers available." },
-    { SIGPOLL, POLL_MSG,     "POLL_MSG",     "Input message available." },
-    { SIGPOLL, POLL_ERR,     "POLL_ERR",     "I/O error." },
-    { SIGPOLL, POLL_PRI,     "POLL_PRI",     "High priority input available." },
-    { SIGPOLL, POLL_HUP,     "POLL_HUP",     "Device disconnected. [Option End]" },
-#endif
-    { -1, -1, NULL, NULL }
-  };
-
-  // Codes valid in any signal context.
-  const struct {
-    int code; const char* s_code; const char* s_desc;
-  } t2 [] = {
-    { SI_USER,      "SI_USER",     "Signal sent by kill()." },
-    { SI_QUEUE,     "SI_QUEUE",    "Signal sent by the sigqueue()." },
-    { SI_TIMER,     "SI_TIMER",    "Signal generated by expiration of a timer set by timer_settime()." },
-    { SI_ASYNCIO,   "SI_ASYNCIO",  "Signal generated by completion of an asynchronous I/O request." },
-    { SI_MESGQ,     "SI_MESGQ",    "Signal generated by arrival of a message on an empty message queue." },
-    // Linux specific
-#ifdef SI_TKILL
-    { SI_TKILL,     "SI_TKILL",    "Signal sent by tkill (pthread_kill)" },
-#endif
-#ifdef SI_DETHREAD
-    { SI_DETHREAD,  "SI_DETHREAD", "Signal sent by execve() killing subsidiary threads" },
-#endif
-#ifdef SI_KERNEL
-    { SI_KERNEL,    "SI_KERNEL",   "Signal sent by kernel." },
-#endif
-#ifdef SI_SIGIO
-    { SI_SIGIO,     "SI_SIGIO",    "Signal sent by queued SIGIO" },
-#endif
-
-#ifdef AIX
-    { SI_UNDEFINED, "SI_UNDEFINED","siginfo contains partial information" },
-    { SI_EMPTY,     "SI_EMPTY",    "siginfo contains no useful information" },
-#endif
-
-#ifdef __sun
-    { SI_NOINFO,    "SI_NOINFO",   "No signal information" },
-    { SI_RCTL,      "SI_RCTL",     "kernel generated signal via rctl action" },
-    { SI_LWP,       "SI_LWP",      "Signal sent via lwp_kill" },
-#endif
-
-    { -1, NULL, NULL }
-  };
-
-  const char* s_code = NULL;
-  const char* s_desc = NULL;
-
-  for (int i = 0; t1[i].sig != -1; i ++) {
-    if (t1[i].sig == si->si_signo && t1[i].code == si->si_code) {
-      s_code = t1[i].s_code;
-      s_desc = t1[i].s_desc;
-      break;
-    }
-  }
-
-  if (s_code == NULL) {
-    for (int i = 0; t2[i].s_code != NULL; i ++) {
-      if (t2[i].code == si->si_code) {
-        s_code = t2[i].s_code;
-        s_desc = t2[i].s_desc;
-      }
-    }
-  }
-
-  if (s_code == NULL) {
-    out->s_name = "unknown";
-    out->s_desc = "unknown";
-    return false;
-  }
-
-  out->s_name = s_code;
-  out->s_desc = s_desc;
-
-  return true;
-}
-
-bool os::signal_sent_by_kill(const void* siginfo) {
-  const siginfo_t* const si = (const siginfo_t*)siginfo;
-  return si->si_code == SI_USER || si->si_code == SI_QUEUE
-#ifdef SI_TKILL
-         || si->si_code == SI_TKILL
-#endif
-  ;
-}
-
-void os::print_siginfo(outputStream* os, const void* si0) {
-
-  const siginfo_t* const si = (const siginfo_t*) si0;
-
-  char buf[20];
-  os->print("siginfo:");
-
-  if (!si) {
-    os->print(" <null>");
-    return;
-  }
-
-  const int sig = si->si_signo;
-
-  os->print(" si_signo: %d (%s)", sig, os::Posix::get_signal_name(sig, buf, sizeof(buf)));
-
-  enum_sigcode_desc_t ed;
-  get_signal_code_description(si, &ed);
-  os->print(", si_code: %d (%s)", si->si_code, ed.s_name);
-
-  if (si->si_errno) {
-    os->print(", si_errno: %d", si->si_errno);
-  }
-
-  // Output additional information depending on the signal code.
-
-  // Note: Many implementations lump si_addr, si_pid, si_uid etc. together as unions,
-  // so it depends on the context which member to use. For synchronous error signals,
-  // we print si_addr, unless the signal was sent by another process or thread, in
-  // which case we print out pid or tid of the sender.
-  if (signal_sent_by_kill(si)) {
-    const pid_t pid = si->si_pid;
-    os->print(", si_pid: %ld", (long) pid);
-    if (IS_VALID_PID(pid)) {
-      const pid_t me = getpid();
-      if (me == pid) {
-        os->print(" (current process)");
-      }
-    } else {
-      os->print(" (invalid)");
-    }
-    os->print(", si_uid: %ld", (long) si->si_uid);
-    if (sig == SIGCHLD) {
-      os->print(", si_status: %d", si->si_status);
-    }
-  } else if (sig == SIGSEGV || sig == SIGBUS || sig == SIGILL ||
-             sig == SIGTRAP || sig == SIGFPE) {
-    os->print(", si_addr: " PTR_FORMAT, p2i(si->si_addr));
-#ifdef SIGPOLL
-  } else if (sig == SIGPOLL) {
-    os->print(", si_band: %ld", si->si_band);
-#endif
-  }
-
-}
-
-bool os::signal_thread(Thread* thread, int sig, const char* reason) {
-  OSThread* osthread = thread->osthread();
-  if (osthread) {
-    int status = pthread_kill(osthread->pthread_id(), sig);
-    if (status == 0) {
-      Events::log(Thread::current(), "sent signal %d to Thread " INTPTR_FORMAT " because %s.",
-                  sig, p2i(thread), reason);
-      return true;
-    }
-  }
-  return false;
-}
-
-int os::Posix::unblock_thread_signal_mask(const sigset_t *set) {
-  return pthread_sigmask(SIG_UNBLOCK, set, NULL);
-}
-
-address os::Posix::ucontext_get_pc(const ucontext_t* ctx) {
-#if defined(AIX)
-   return Aix::ucontext_get_pc(ctx);
-#elif defined(BSD)
-   return Bsd::ucontext_get_pc(ctx);
-#elif defined(LINUX)
-   return Linux::ucontext_get_pc(ctx);
-#else
-   VMError::report_and_die("unimplemented ucontext_get_pc");
-#endif
-}
-
-void os::Posix::ucontext_set_pc(ucontext_t* ctx, address pc) {
-#if defined(AIX)
-   Aix::ucontext_set_pc(ctx, pc);
-#elif defined(BSD)
-   Bsd::ucontext_set_pc(ctx, pc);
-#elif defined(LINUX)
-   Linux::ucontext_set_pc(ctx, pc);
-#else
-   VMError::report_and_die("unimplemented ucontext_get_pc");
-#endif
-}
-
 char* os::Posix::describe_pthread_attr(char* buf, size_t buflen, const pthread_attr_t* attr) {
   size_t stack_size = 0;
   size_t guard_size = 0;
@@ -1608,22 +1004,6 @@ static void pthread_init_common(void) {
   os::PlatformMutex::init();
 }
 
-sigset_t sigs;
-struct sigaction sigact[NSIG];
-
-struct sigaction* os::Posix::get_preinstalled_handler(int sig) {
-  if (sigismember(&sigs, sig)) {
-    return &sigact[sig];
-  }
-  return NULL;
-}
-
-void os::Posix::save_preinstalled_handler(int sig, struct sigaction& oldAct) {
-  assert(sig > 0 && sig < NSIG, "vm signal out of expected range");
-  sigact[sig] = oldAct;
-  sigaddset(&sigs, sig);
-}
-
 // Not all POSIX types and API's are available on all notionally "posix"
 // platforms. If we have build-time support then we will check for actual
 // runtime support via dlopen/dlsym lookup. This allows for running on an
@@ -1726,7 +1106,6 @@ void os::Posix::init_2(void) {
                (_pthread_condattr_setclock != NULL ? "" : " not"));
   log_info(os)("Relative timed-wait using pthread_cond_timedwait is associated with %s",
                _use_clock_monotonic_condattr ? "CLOCK_MONOTONIC" : "the default clock");
-  sigemptyset(&sigs);
 }
 
 #else // !SUPPORTS_CLOCK_MONOTONIC
@@ -1739,7 +1118,6 @@ void os::Posix::init_2(void) {
   log_info(os)("Use of CLOCK_MONOTONIC is not supported");
   log_info(os)("Use of pthread_condattr_setclock is not supported");
   log_info(os)("Relative timed-wait using pthread_cond_timedwait is associated with the default clock");
-  sigemptyset(&sigs);
 }
 
 #endif // SUPPORTS_CLOCK_MONOTONIC
